+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2018-11-23
+### Added
+
 ## [0.1.2] - 2018-11-23
 ### Added
 - `SQLiteDatabase#inTransaction` getter
@@ -46,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SQLiteDatabase#rawQuery()` method
 - `SQLiteDatabase#setLocale()` method
 
+[0.1.3]:  https://github.com/drydart/flutter_sqlcipher/compare/0.1.2...0.1.3
 [0.1.2]:  https://github.com/drydart/flutter_sqlcipher/compare/0.1.1...0.1.2
 [0.1.1]:  https://github.com/drydart/flutter_sqlcipher/compare/0.1.0...0.1.1
 [0.1.0]:  https://github.com/drydart/flutter_sqlcipher/compare/0.0.6...0.1.0